requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras-nightly==3.7.0.dev2024122803
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
sympy==1.13.1
tb-nightly==2.19.0a20241228
tensorboard-data-server==0.7.2
termcolor==2.5.0
tf-nightly==2.19.0.dev20241217
tf_nightly_intel==2.19.0.dev20241217
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0